Removing monads. Fixes #884 (#886)
* Removing monads. Fixes #884

* Fixing post fetching.

* Dont show not_logged_in error for navbar.

* Adding the lemmy-js-client RC.

* Fixing registration application mode

+++ src/shared/components/community/sidebar.tsx
@@ -1,4 +1,3 @@
-import { None, Option, Some } from "@sniptt/monads";
 import { Component, linkEvent } from "inferno";
 import { Link } from "inferno-router";
 import {
@@ -13,7 +12,6 @@ import {
   PurgeCommunity,
   RemoveCommunity,
   SubscribedType,
-  toUndefined,
 } from "lemmy-js-client";
 import { i18n } from "../../i18next";
 import { UserService, WebSocketService } from "../../services";
@@ -21,9 +19,9 @@ import {
   amAdmin,
   amMod,
   amTopMod,
-  auth,
   getUnixTime,
   mdToHtml,
+  myAuth,
   numToSI,
   wsClient,
 } from "../../utils";
@@ -39,7 +37,7 @@ interface SidebarProps {
   admins: PersonViewSafe[];
   allLanguages: Language[];
   siteLanguages: number[];
-  communityLanguages: Option<number[]>;
+  communityLanguages?: number[];
   online: number;
   enableNsfw?: boolean;
   showIcon?: boolean;
@@ -47,31 +45,27 @@ interface SidebarProps {
 }
 
 interface SidebarState {
-  removeReason: Option<string>;
-  removeExpires: Option<string>;
+  removeReason?: string;
+  removeExpires?: string;
   showEdit: boolean;
   showRemoveDialog: boolean;
   showPurgeDialog: boolean;
-  purgeReason: Option<string>;
+  purgeReason?: string;
   purgeLoading: boolean;
   showConfirmLeaveModTeam: boolean;
 }
 
 export class Sidebar extends Component<SidebarProps, SidebarState> {
-  private emptyState: SidebarState = {
+  state: SidebarState = {
     showEdit: false,
     showRemoveDialog: false,
-    removeReason: None,
-    removeExpires: None,
     showPurgeDialog: false,
-    purgeReason: None,
     purgeLoading: false,
     showConfirmLeaveModTeam: false,
   };
 
   constructor(props: any, context: any) {
     super(props, context);
-    this.state = this.emptyState;
     this.handleEditCommunity = this.handleEditCommunity.bind(this);
     this.handleEditCancel = this.handleEditCancel.bind(this);
   }
@@ -83,7 +77,7 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
           this.sidebar()
         ) : (
           <CommunityForm
-            community_view={Some(this.props.community_view)}
+            community_view={this.props.community_view}
             allLanguages={this.props.allLanguages}
             siteLanguages={this.props.siteLanguages}
             communityLanguages={this.props.communityLanguages}
@@ -336,13 +330,12 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
   }
 
   description() {
-    let description = this.props.community_view.community.description;
-    return description.match({
-      some: desc => (
+    let desc = this.props.community_view.community.description;
+    return (
+      desc && (
         <div className="md-div" dangerouslySetInnerHTML={mdToHtml(desc)} />
-      ),
-      none: <></>,
-    });
+      )
+    );
   }
 
   adminButtons() {
@@ -350,7 +343,7 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
     return (
       <>
         <ul className="list-inline mb-1 text-muted font-weight-bold">
-          {amMod(Some(this.props.moderators)) && (
+          {amMod(this.props.moderators) && (
             <>
               <li className="list-inline-item-action">
                 <button
@@ -362,7 +355,7 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
                   <Icon icon="edit" classes="icon-inline" />
                 </button>
               </li>
-              {!amTopMod(Some(this.props.moderators)) &&
+              {!amTopMod(this.props.moderators) &&
                 (!this.state.showConfirmLeaveModTeam ? (
                   <li className="list-inline-item-action">
                     <button
@@ -401,7 +394,7 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
                     </li>
                   </>
                 ))}
-              {amTopMod(Some(this.props.moderators)) && (
+              {amTopMod(this.props.moderators) && (
                 <li className="list-inline-item-action">
                   <button
                     className="btn btn-link text-muted d-inline-block"
@@ -466,7 +459,7 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
                 id="remove-reason"
                 className="form-control mr-2"
                 placeholder={i18n.t("optional")}
-                value={toUndefined(this.state.removeReason)}
+                value={this.state.removeReason}
                 onInput={linkEvent(this, this.handleModRemoveReasonChange)}
               />
             </div>
@@ -496,7 +489,7 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
                 id="purge-reason"
                 className="form-control mr-2"
                 placeholder={i18n.t("reason")}
-                value={toUndefined(this.state.purgeReason)}
+                value={this.state.purgeReason}
                 onInput={linkEvent(this, this.handlePurgeReasonChange)}
               />
             </div>
@@ -533,12 +526,15 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
 
   handleDeleteClick(i: Sidebar, event: any) {
     event.preventDefault();
-    let deleteForm = new DeleteCommunity({
-      community_id: i.props.community_view.community.id,
-      deleted: !i.props.community_view.community.deleted,
-      auth: auth().unwrap(),
-    });
-    WebSocketService.Instance.send(wsClient.deleteCommunity(deleteForm));
+    let auth = myAuth();
+    if (auth) {
+      let deleteForm: DeleteCommunity = {
+        community_id: i.props.community_view.community.id,
+        deleted: !i.props.community_view.community.deleted,
+        auth,
+      };
+      WebSocketService.Instance.send(wsClient.deleteCommunity(deleteForm));
+    }
   }
 
   handleShowConfirmLeaveModTeamClick(i: Sidebar) {
@@ -546,19 +542,18 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
   }
 
   handleLeaveModTeamClick(i: Sidebar) {
-    UserService.Instance.myUserInfo.match({
-      some: mui => {
-        let form = new AddModToCommunity({
-          person_id: mui.local_user_view.person.id,
-          community_id: i.props.community_view.community.id,
-          added: false,
-          auth: auth().unwrap(),
-        });
-        WebSocketService.Instance.send(wsClient.addModToCommunity(form));
-        i.setState({ showConfirmLeaveModTeam: false });
-      },
-      none: void 0,
-    });
+    let mui = UserService.Instance.myUserInfo;
+    let auth = myAuth();
+    if (auth && mui) {
+      let form: AddModToCommunity = {
+        person_id: mui.local_user_view.person.id,
+        community_id: i.props.community_view.community.id,
+        added: false,
+        auth,
+      };
+      WebSocketService.Instance.send(wsClient.addModToCommunity(form));
+      i.setState({ showConfirmLeaveModTeam: false });
+    }
   }
 
   handleCancelLeaveModTeamClick(i: Sidebar) {
@@ -568,46 +563,50 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
   handleUnsubscribe(i: Sidebar, event: any) {
     event.preventDefault();
     let community_id = i.props.community_view.community.id;
-    let form = new FollowCommunity({
-      community_id,
-      follow: false,
-      auth: auth().unwrap(),
-    });
-    WebSocketService.Instance.send(wsClient.followCommunity(form));
+    let auth = myAuth();
+    if (auth) {
+      let form: FollowCommunity = {
+        community_id,
+        follow: false,
+        auth,
+      };
+      WebSocketService.Instance.send(wsClient.followCommunity(form));
+    }
 
     // Update myUserInfo
-    UserService.Instance.myUserInfo.match({
-      some: mui =>
-        (mui.follows = mui.follows.filter(i => i.community.id != community_id)),
-      none: void 0,
-    });
+    let mui = UserService.Instance.myUserInfo;
+    if (mui) {
+      mui.follows = mui.follows.filter(i => i.community.id != community_id);
+    }
   }
 
   handleSubscribe(i: Sidebar, event: any) {
     event.preventDefault();
     let community_id = i.props.community_view.community.id;
-    let form = new FollowCommunity({
-      community_id,
-      follow: true,
-      auth: auth().unwrap(),
-    });
-    WebSocketService.Instance.send(wsClient.followCommunity(form));
+    let auth = myAuth();
+    if (auth) {
+      let form: FollowCommunity = {
+        community_id,
+        follow: true,
+        auth,
+      };
+      WebSocketService.Instance.send(wsClient.followCommunity(form));
+    }
 
     // Update myUserInfo
-    UserService.Instance.myUserInfo.match({
-      some: mui =>
-        mui.follows.push({
-          community: i.props.community_view.community,
-          follower: mui.local_user_view.person,
-        }),
-      none: void 0,
-    });
+    let mui = UserService.Instance.myUserInfo;
+    if (mui) {
+      mui.follows.push({
+        community: i.props.community_view.community,
+        follower: mui.local_user_view.person,
+      });
+    }
   }
 
   get canPost(): boolean {
     return (
       !this.props.community_view.community.posting_restricted_to_mods ||
-      amMod(Some(this.props.moderators)) ||
+      amMod(this.props.moderators) ||
       amAdmin()
     );
   }
@@ -617,25 +616,28 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
   }
 
   handleModRemoveReasonChange(i: Sidebar, event: any) {
-    i.setState({ removeReason: Some(event.target.value) });
+    i.setState({ removeReason: event.target.value });
   }
 
   handleModRemoveExpiresChange(i: Sidebar, event: any) {
-    i.setState({ removeExpires: Some(event.target.value) });
+    i.setState({ removeExpires: event.target.value });
   }
 
   handleModRemoveSubmit(i: Sidebar, event: any) {
     event.preventDefault();
-    let removeForm = new RemoveCommunity({
-      community_id: i.props.community_view.community.id,
-      removed: !i.props.community_view.community.removed,
-      reason: i.state.removeReason,
-      expires: i.state.removeExpires.map(getUnixTime),
-      auth: auth().unwrap(),
-    });
-    WebSocketService.Instance.send(wsClient.removeCommunity(removeForm));
+    let auth = myAuth();
+    if (auth) {
+      let removeForm: RemoveCommunity = {
+        community_id: i.props.community_view.community.id,
+        removed: !i.props.community_view.community.removed,
+        reason: i.state.removeReason,
+        expires: getUnixTime(i.state.removeExpires),
+        auth,
+      };
+      WebSocketService.Instance.send(wsClient.removeCommunity(removeForm));
 
-    i.setState({ showRemoveDialog: false });
+      i.setState({ showRemoveDialog: false });
+    }
   }
 
   handlePurgeCommunityShow(i: Sidebar) {
@@ -643,39 +645,51 @@ export class Sidebar extends Component<SidebarProps, SidebarState> {
   }
 
   handlePurgeReasonChange(i: Sidebar, event: any) {
-    i.setState({ purgeReason: Some(event.target.value) });
+    i.setState({ purgeReason: event.target.value });
   }
 
   handlePurgeSubmit(i: Sidebar, event: any) {
     event.preventDefault();
 
-    let form = new PurgeCommunity({
-      community_id: i.props.community_view.community.id,
-      reason: i.state.purgeReason,
-      auth: auth().unwrap(),
-    });
-    WebSocketService.Instance.send(wsClient.purgeCommunity(form));
-
-    i.setState({ purgeLoading: true });
+    let auth = myAuth();
+    if (auth) {
+      let form: PurgeCommunity = {
+        community_id: i.props.community_view.community.id,
+        reason: i.state.purgeReason,
+        auth,
+      };
+      WebSocketService.Instance.send(wsClient.purgeCommunity(form));
+      i.setState({ purgeLoading: true });
+    }
   }
 
   handleBlock(i: Sidebar, event: any) {
     event.preventDefault();
-    let blockCommunityForm = new BlockCommunity({
-      community_id: i.props.community_view.community.id,
-      block: true,
-      auth: auth().unwrap(),
-    });
-    WebSocketService.Instance.send(wsClient.blockCommunity(blockCommunityForm));
+    let auth = myAuth();
+    if (auth) {
+      let blockCommunityForm: BlockCommunity = {
+        community_id: i.props.community_view.community.id,
+        block: true,
+        auth,
+      };
+      WebSocketService.Instance.send(
+        wsClient.blockCommunity(blockCommunityForm)
+      );
+    }
   }
 
   handleUnblock(i: Sidebar, event: any) {
     event.preventDefault();
-    let blockCommunityForm = new BlockCommunity({
-      community_id: i.props.community_view.community.id,
-      block: false,
-      auth: auth().unwrap(),
-    });
-    WebSocketService.Instance.send(wsClient.blockCommunity(blockCommunityForm));
+    let auth = myAuth();
+    if (auth) {
+      let blockCommunityForm: BlockCommunity = {
+        community_id: i.props.community_view.community.id,
+        block: false,
+        auth,
+      };
+      WebSocketService.Instance.send(
+        wsClient.blockCommunity(blockCommunityForm)
+      );
+    }
   }
 }
